Updated DashboardHeader to be the same across dashboards.
