+++ src/components/UserDashboardContent.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import { connect } from "react-redux";
-import { useHistory } from "react-router-dom";
 import TeamList from './UserTeamsList';
 import { Layout, Tooltip, Avatar, Typography, Card, Icon, Button } from 'antd';
-
-import { logoutUser } from "../redux/actions/userActions";
+import DashboardHeader from './DashboardHeader';
 
 import UserVideos from "./UserVideosList"
 const { Title } = Typography;
@@ -12,29 +10,10 @@ const { Title } = Typography;
 const { Header, Content } = Layout;
 
 function UserDashboardContent(props) {
-	let history = useHistory();
-
-	const handleLogout = (e) => {
-		e.preventDefault();
-		props.logoutUser();
-		history.push("/");
-	}
 
 	return (
 		<Layout>
-			{/* user dash content area */}
-			<Header className="userDashHeader">
-				<div className="userDashContentHeader">
-					<Title level={4}>{props.username}</Title>
-					<Tooltip placement="left" title="username here">
-						{/* change src below for image */}
-						<Avatar size="large" icon="user" src="" />
-					</Tooltip>
-				</div>
-			<Button onClick={handleLogout}>
-				<Icon type="logout" style={{ fontSize: '32px' }} />
-			</Button>
-			</Header>
+			<DashboardHeader />
 			<Card title="Your Teams" style={{ margin: "20px" }}>
 				<TeamList />
 			</Card>
@@ -51,8 +30,4 @@ const mapStateToProps = (state) => {
 	}
 }
 
-const mapActionsToProps = {
-  logoutUser
-};
-
-export default connect(mapStateToProps, mapActionsToProps)(UserDashboardContent);
+export default connect(mapStateToProps, {})(UserDashboardContent);
